+++ src/test/suite/extension.test.ts
@@ -1,9 +1,6 @@
 import { strictEqual } from 'assert';
-import { window } from 'vscode';
 
-suite('Extension Test Suite', async () => {
-	await window.showInformationMessage('Start all tests.');
-
+suite('Extension test', () => {
 	test('Sample test', () => {
 		strictEqual(-1, [1, 2, 3].indexOf(5));
 		strictEqual(-1, [1, 2, 3].indexOf(0));